UI, leaders board, timer and competitive challenge
The idea for this commit was to add a goal to the game. Now the timer is ticking back so there is a time limit in which we need to make as much points as possible.
Also added this UI/UX:
- Death screen (with flexible death reasons)
- Welcome screen asking for a name
- Winners screen (can see when the 30 seconds pass) with the leaders chart

Now the game has a goal - perform better than your friends in 30 seconds